fix the vlm connector filename
