+++ plot_logits.py
@@ -15,8 +15,8 @@ from transformers import GPT2Tokenizer
 from PIL import Image
 from gpt import GPT, GPTConfig, transpose_specific_layers
 from clip import load_clip
+from vision_language_connector_bad import VisionLanguageConnectorBad
 from vision_language_connector import VisionLanguageConnector
-from vision_language_connector_old import VisionLanguageConnectorOld
 
 os.environ["PYTORCH_CUDA_ALLOC_CONF"] = "max_split_size_mb:128"
 
@@ -291,12 +291,12 @@ if __name__ == "__main__":
     coco_ann_file = "../coco/annotations/captions_val2017.json"
     device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 
-    connector_weights_path = "vl_connector_large_mlp_collapsed.pt"
-    # connector_weights_path = "vl_connector.pt"
+    # connector_weights_path = "vl_connector_large_mlp_collapsed.pt"
+    connector_weights_path = "vl_connector.pt"
 
     vision_encoder, preprocess = load_clip(device)
+    # connector = VisionLanguageConnectorBad()
     connector = VisionLanguageConnector()
-    # connector = VisionLanguageConnectorOld()
     connector.load_state_dict(torch.load(connector_weights_path, map_location="cpu"))
 
     connector = connector.to(device)
